Chatter bot Research Closes #1
Working on how to design a chatter bot i read mark Humprys paper on his
chatter bot MGonz.(unsure if aloud to directly link on a public
repository, will ask and if so link in the planning readme) PS write the
readme for the planning folder...

Marks chatter bot is written in lisp because iam unfamiliar with lisp i
decided to port it to php. in doing so would also give me a better
understanding of the chatter bots design. After porting to php i found i
could very eaisly using code i wrote from my 3rd year project connect
the bot to twitter(This is currently untested) the twitterr port uses
the codebird library to connect to the twitter api.

After working on this i have decided to do the neural network project as
i was getting more excited about his idea and is shaping up to be a fun
project to work on.