New Input Scheme.. Why net failing
Changed the input scheme to only be the tiles in a nine tile radius with
the enemy on screen represented as a -1 in the tile array. Figured out
why it isnt jumping at a enemy see file "why_it_cant_jump" line 185 -
187 exemplars it is being trained on look to be getting the inputs after
the button is pressed so it never shows the orginal frame that should
trigger the output. it will never learn to jump as it can never meet the
conditions to jump in the net with out jumping first! Should be a simple
fix just need to move around function call's or else create a delay of
some kind..