XML for logs and net values. cleaned up file struct
Changed the log functions to write to xml files. much better to read
then the dat files and will make debuging the neat and experimentation
much easier. will look at writing a style sheet for the xml to display
the data much nicer. also changed the files structer for my data files
removed all the old ones in the DAT_File folder and added seperate
folders to catigories logs just to make it all easier and start from
fresh as it was getting seriously blothed that folder!