styling on XML
Wrote two quick little xsl files for the run and training logs. So much
better to read. Didnt do one for the network values as those will
generaly not be read will just be loaded into the system. TODO: write a
function to parse this xml file and load the network values.