Seperated Learning & Pretty display of Net
Seperated the trainning from bizhawk and added a config varaible to
allow the user to pass the network values file name into bizhawk.
Also implemented a nice little display of the network removed the (-1,1)
display of net to a more colourful one and made a little controler
display for the outputs.