Pre-Processing and Training Analysis
Wrote a java file to pre-process the Exemplar files.
Also added to the Learn function in NeuralNet.lua to record the total
Error rate at each iteration of the exemplars to a csv.
Along with this added a r file to graph the error rate over the
iterations... helps with seeing if any changes i do to the net are
actualy taking effect so far the pre-processing is greatly improving the
error :)