Reduce radius function
Wrote a java file to reduce an exemplar file to a smaller radius. Going
to go about recording exemplars at the max radius and the just reducing
the radius if needs be. much better then when i want to reduce the
inputs having to re record at a smaller radius and try to match the
original gameplay....should have really done this first.