q-learning implemented
q-learning is impmented but i dont think how im back propigating the
error is working correctly as if it keeps choosing the same action that
has a negative reinforcement value it will keep ruducing the qvalue but
will stop at 0.15625(Everytime) and not go under it resulting in some
actions not being called as this is the highest value. Now this looks to
be a issue with my reinforcment values as at the moment i just reward
the agent when it goes further right in the level how ever it appears
that that if the agent chooses right takes the action that player
actually does not move right and thus not increasing its x value? After
steping through how mario moves it actually takes 5 frames for mario to
increase its x value by 1. As i call actions on a frame by frame basis
it will be penilised 4 times untill it gets the correct value even
though it is doing some thing write.