Fix in gathering new state y after taking action a
When i first implemnted q-learning i saw that there might be a issue in
that a change in state is not reflected by a change in Frame. i thought
it might be ok to leave it and just do the q-learning on a frame by
frame basis, it wasnt going anywhere. I implemnted a fix that basically
says when the agent takes a action keep advancing frames untill it
reaches a different input then the one previous. This worked a treat and
it is starting to fly along. I just have to get the tempature in my
boltzman distrobution for selecting the action to increase as the
training goes one.... Ie so that in the early stages of traing i want it
to choose exploration over explotation and then in the later statges
choose explotation over exploratoin. Really wish i had of went straight
with q-learning from the begining such a fun technique and havnt even
started designing the rewards in great detail.