Function to detect getting over an obsticle
Can now detect if an action taken when the agent encounters and obstilce
actually results in the agent getting closer to get over it.
TODO: need to do the same but for an enemy, should be quick have all the
componts for it done with this function.
TODO2: detect when they have gotten over... again should be fine just
using the same method i can pretty easily detect if the new state puts a
tile under mario... m[VIEW_RADIUS + 3][VIEW_RADIUS + 1] == 1

For the enemy just detect if the enemy is in any of the columns to the
left of column VEIW_RADIUS + 1