Removed the mgonz code
Removed the Mgonz code just as it was anoying me how it affected the
code breakdown of the repository and does absolutly noting in terms of
the project. except that i wrote it in my first week of college to
decide if i was going to make a twitter chatter bot or this.