Finished the first level with supervised learning
Got a succesfull network trained on supervised learning to beat the
first level.
Also added a couple of fixes to q-learning and found a very big issue
that i will log.
created a folder called documentation just to clean everything up a bit
and added a video of the network playing the first level which i will
add to my blog.