Experience Replay and Fixed issue from last commit
-Added backward experience replay to the network.

-Added 2 new actions to the network due to the issue as mentioned in the
last commit that for some obstecals in mario the agent has to press jump
with either the left or right key pressed also at the same time.  It is
giving very promising results as it made it very far into the level,
alought better then expected as i thought it would have issue with the
pits in the level as did the supervised network but it seemed to get the
hang off them very quickly.

-Added back in the configuration of 1 network and sending a actions in
as inputs but this does not do very well as the network has a very hard
time distingushing the inputs and actions, left it just to have in case
i want to go back and explore this configuration again.

There is a bug that has been hidding for a while that for some reason
the qxa variable gets set to null. finding it diffecult to locate the
origin as this could happen at a number of code breaks. Also it appears
to be random when this occurs, Although noting is random i probably just
cant find the pattern that is causing it. Will need to dump the values
of this variable and the variables that give it its value to a file so i
can hopefully find the cause of it as going over the code doesnt seem to
indicate any issue.
This bug should be the highest priority before i do anymore
modifications to the network.