Fixed Issue where qxa was getting set to null
Fixed the issue where qxa was getting set to null
This was due to the randomFloat function returning
returning zero because i was chosing a range between
zero and one. my mistake fixed now by selecting a
random value between 0.1 and 1. I knew the bug was
random :P

Begun training and it works very well the problem i can see is that it
might never actually reach the end of the level if i train it for a set
number of interactions so im going to modify it to train as follows:
-interact with the level untill it completes it
-replay experiences from run
-repeat for x number of runs

Another thing added was a function to detect the level as a really funny
thing happened during training. It found a shortcut in the level(I DIDNT
EVEN KNOW EXISTED!!) where on the third pipe you can go down it and come
out at a pipe right at the end. It did this made it to the end of the
level and beat it in ~2000 epochs, the training was for 5000 epochs.
because of this when it reached the end of the level the game reduces
the time limit to zero and adds it to the score but i have a fucntion
that says if the timer reaches bellow 200 set it to 999 so that mario
can interact with the world all it likes. This caused a infanite loop
where mario was no longer able to move and the alogrithm kept running
geting a negative reward for any action it perfromed.