Added terminal state check and change reinforce
Added fucntion so that if the state is a terminal state to just set ok
to be the reinforcement value for that state.

Updated the reinforcement values so that if it is a terminal to just
return the R value for the terminal condition and not other
reinforcments .

Started training a net with these changes and with a extra 20 nuerons
and seems to be performing much better, forgeting is not as large of a
issue but is still slightly present with the first enemy.